Re-enable and fix ApplicationTest
