+++ api/src/integration/java/org/pmiops/workbench/ApplicationTest.java
@@ -4,7 +4,6 @@ import static com.google.common.truth.Truth.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,7 +15,6 @@ import org.springframework.stereotype.Component;
  * This test that all application injection is done properly. It loads all JPA repositories,
  * Services, Controllers, Components and Configurations.
  */
-@Disabled
 @SpringBootTest
 public class ApplicationTest {
 
@@ -30,8 +28,6 @@ public class ApplicationTest {
    * of spring.datasource.platform. This allows you to switch to database-specific scripts if
    * necessary. For example, you might choose to set it to the vendor name of the database (hsqldb,
    * h2, oracle, mysql, postgresql, and so on).
-   *
-   * <p>Ignoring this test due to observed flakiness and false positives. RW-4806
    */
   @Test
   public void contextLoads() {
